#40 commenting feature is now functional, just need to fix how it looks
